final fix

+++ project/orders/views.py
@@ -15,8 +15,6 @@ logger = logging.getLogger(__name__)
 
 @login_required()
 def create_order(request):
-    is_create_order_url = request.path.endswith('create_order/')
-
     if request.method == 'POST':
         form = CreateOrderForm(data=request.POST)
         if form.is_valid():
@@ -24,23 +22,25 @@ def create_order(request):
                 with transaction.atomic():
                     user = request.user
                     cart_items = Cart.objects.filter(user=user)
-
                     if cart_items.exists():
                         order = Order.objects.create(
                             user=user,
                             phone_number=form.cleaned_data['phone_number'],
                             requires_delivery=form.cleaned_data['requires_delivery'],
-                            delivery_address=form.cleaned_data['delivery_address'] if form.cleaned_data['requires_delivery'] else '',
+                            delivery_address=form.cleaned_data['delivery_address'] if form.cleaned_data['requires_delivery'] == '1' else '',
                             payment_on_get=form.cleaned_data['payment_on_get'],
                         )
+                        total_price = 0
                         for cart_item in cart_items:
                             product = cart_item.product
                             name = cart_item.product.name
-                            price = cart_item.product.sell_price()
+                            price = product.sell_price()
                             quantity = cart_item.quantity
 
-                            if product.quantity < quantity:
-                                raise ValidationError(f'Insufficient quantity of goods {name} on stock. In stock - {product.quantity}')
+                            item_total_price = round(price * quantity, 2)
+                            if item_total_price > 99999999.99:
+                                logger.error(f'Item total price exceeds limit: {name} with total price {item_total_price}')
+                                raise ValidationError(f'Total price for item {name} exceeds the limit.')
 
                             OrderItem.objects.create(
                                 order=order,
@@ -49,8 +49,14 @@ def create_order(request):
                                 price=price,
                                 quantity=quantity,
                             )
-                            product.quantity -= quantity
+                            # product.quantity -= quantity
                             product.save()
+
+                            total_price += item_total_price
+
+                        if total_price > 99999999.99:
+                            logger.error(f'Total order price exceeds limit: {total_price}')
+                            raise ValidationError(f'Total order price exceeds the limit.')
 
                         cart_items.delete()
                         return redirect('accounts:profile')
@@ -68,8 +74,8 @@ def create_order(request):
         form = CreateOrderForm(initial=initial)
 
     context = {
-        'is_create_order_url': is_create_order_url,
         'title': 'Order placement',
         'form': form,
     }
     return render(request, 'orders/create_order.html', context=context)
+
